# booking-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking Billiard API",
        "description": "Welcome to the official **Booking Billiard API** documentation.\n\nThis API provides an enterprise-grade backend infrastructure designed to streamline and automate billiard center operations. Built with a focus on high performance, security, and scalability, it serves as the core engine for modern booking applications.\n\n### Core Capabilities:\n- **Secure Authentication**: Robust role-based access control (RBAC) powered by Laravel Sanctum, ensuring data privacy and operational security.\n- **Resource Management**: Dynamic and highly configurable table allocation, capacity planning, and scheduling algorithms.\n- **Advanced Reservation System**: Sophisticated booking engine capable of handling future reservations, real-time availability conflict resolution, and instant walk-in assignments.\n- **Transaction Architecture**: Seamless and transactional payment state management, ensuring financial data integrity.\n- **Analytics & Telemetry**: Comprehensive real-time monitoring of table states and data-driven reporting for business intelligence.\n\n---\n\n*Architected and developed by **Wisnu Nugraha**.*",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/reports/bestsellers": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get most booked tables (Admin)",
                "operationId": "a2f646f723c82ca701afea2f55b5e905",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "table": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string"
                                                    },
                                                    "total_bookings": {
                                                        "type": "integer"
                                                    },
                                                    "total_revenue": {
                                                        "type": "number"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/busiest-schedules": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get busiest schedules (Admin)",
                "operationId": "e79159f327d8991cf322569586b095a9",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "day_name": {
                                                        "type": "string"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "format": "time"
                                                    },
                                                    "total_bookings": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}